requirements.txt
click==7.1.2
dataclasses==0.6
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
logger==1.4
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.5
PySnooper==0.4.1
PyYAML==5.3.1
Werkzeug==1.0.1
